Rename graalEntry.ts to spawnRendererEntry.ts and document its consumer
